Extract VariableNamingSpec from NamingRulesSpec
